Added role based navBar
After this update, once the user logs out, the users cannot access various endpoints of templates.
The configure option will be hidden for TA roles.
If the user tries to break in to any enpoint without login, he will be redirected to the welcome screen.
In the next commit, the user will be redirected to a page not found endpoint rather than welcome screen